==> notebooks/1.0-initial-data-exploration.ipynb <==
# packages to store and manipulate data
import pandas as pd
import numpy as np

# import data
df_training = pd.read_csv('../data/raw/task1_training.tsv', sep='\t')
df_validation = pd.read_csv('../data/raw/task1_validation.tsv', sep='\t')
df_test = pd.read_csv('../data/raw/AnnotationDRUGSInTweets_EMNLPChallenge18_TrainingSetClean.tsv', sep='\t')

# check sample of training data
df_training.sample(5)

# check sample of validation data
df_validation.sample(5)

# check sample of test data
df_test.sample(5)

df_training['class'].value_counts()

df_training.loc[df_training['class'] == 0].sample(5)[['tweet']]

df_validation['class'].value_counts()

df_test['label'].value_counts()

# rename columns
df_training = df_training.rename(columns={'class' : 'label', 'tweet': 'text'})
df_validation = df_validation.rename(columns={'class' : 'label', 'tweet': 'text'})

#### Export data with necessary columns only

df_training.to_csv('../data/interim/train.csv', columns = ['label', 'text'], index=False, header=False)
df_validation.to_csv('../data/interim/valid.csv', columns = ['label', 'text'], index=False, header=False)
df_test.to_csv('../data/interim/test.csv', columns = ['label', 'text'], index=False, header=False)
