==> notebooks/1.2-Pytorch-Dataset.ipynb <==
import torch
from torch.utils.data import Dataset

import pandas as pd
import numpy as np

from transformers import BertTokenizer
tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

class MedsDataset(Dataset):
    """Medications mention dataset"""

    def __init__(self, data_path, max_len):
        """
        Args:
            data_path (string): Path to the file containg texts with labels.
        """
        self.df = pd.read_csv(data_path, header=None)
        self.MAX_LEN = max_len

    def __len__(self):
        return len(self.df)

    def __getitem__(self, idx):
        """
            Output:
            text: 
            attention_mask: 
            label: 
        """
        label = self.df.iloc[idx, 0]
        text = self.df.iloc[idx, 1]
        tokens = tokenizer.encode(text, add_special_tokens=True)
        padded = np.array(tokens + [0]*(self.MAX_LEN-len(tokens)))
        text = torch.tensor(padded)
        attention_mask = np.where(padded != 0, 1, 0)
        attention_mask = torch.tensor(attention_mask)
        return (label, text, attention_mask)

train = MedsDataset('../data/processed/train.csv', 128)
valid = MedsDataset('../data/processed/valid.csv', 128)

train[1016]

df_train = pd.read_csv('../data/processed/train.csv', header=None)

df_train.shape

text = df_train.iloc[1016, 1]

text

tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

tokens = tokenizer.encode(text, add_special_tokens=True)

tokens

label = df_train.iloc[0, 0]

label
