==> notebooks/6.0-Create-Dataset-Class.ipynb <==
## Create vocabulary

with open('../src/models/.vector_cache/glove.6B.300d.txt') as f:
    words = f.readlines()

len(words)

word2id = dict()

embeddings = []

for i, w in enumerate(words):
    w_list = w.split(' ')
    word2id.update({w_list[0]: i})
    embeddings.append([float(j) for j in w_list[1:]])

words[3]

word2id['of']

embeddings[3]

import torch
import io

from torch.utils.data import DataLoader

from torchtext.data.utils import ngrams_iterator
from torchtext.data.utils import get_tokenizer
from torchtext.utils import unicode_csv_reader

data_path = '../data/processed/train.csv'

tokenizer = get_tokenizer("basic_english")

with io.open(data_path, encoding="utf8") as f:
    reader = unicode_csv_reader(f)
    import pdb; pdb.set_trace()
    for row in reader:
        continue
