==> notebooks/7.0-Task-Prediction.ipynb <==
import torch
import torch.nn as nn
import torch.nn.functional as F
from torch.utils.data import Dataset
from torch.utils.data import DataLoader

import pandas as pd
import numpy as np

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

df = pd.read_csv('../data/raw/task1_test_participant.tsv', sep='\t')

df.sample(5)

df = df[['text', 'tweet_id']]

import os
os.getcwd()

os.chdir('/home/zqxh49/Development/phd/meds-classifier/')
from model.model import BERTGRUModel

model = torch.load('saved/models/train/model.bin')
model.eval()
model = model.to(device)

from transformers import BertTokenizer
tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

max_seq_length = 128
class MedsDataset(Dataset):
    """Medications mention dataset"""

    def __init__(self, data, num_lines=None):
        """
        Args:
            data (string): Path to the file containg texts with labels.
        """
#         self.df = pd.read_csv(data_path, header=None)
        self.data = data
        self.max_seq_length = max_seq_length
        if num_lines:
            self.data = self.data[:num_lines]

    def __len__(self):
        return len(self.data)

    def __getitem__(self, idx):
        """
            Output:
            text: 
            attention_mask: 
            tweet_id: 
        """
        tweet_id = self.data.iloc[idx, 1]
        text = self.data.iloc[idx, 0]
        tokens = tokenizer.encode(text, add_special_tokens=True)
        padded = np.array(tokens + [0]*(max_seq_length-len(tokens)))
        text = torch.tensor(padded)
        attention_mask = np.where(padded != 0, 1, 0)
        attention_mask = torch.tensor(attention_mask)
        return text, attention_mask, tweet_id

text_dataset = MedsDataset(df)

text_dataset[0]

batch_size = 16
num_workers = 3
data = DataLoader(
        text_dataset,
        batch_size=batch_size,
        num_workers=num_workers)

from tqdm import tqdm

y_pred = {}
with tqdm(unit_scale=0, unit='lines', total=len(text_dataset)) as t:
    for i, (text, attention_mask, tweet_id) in enumerate(data):
        t.update(len(text))
        with torch.no_grad():
            input_ids, input_mask = text.to(device), attention_mask.to(device)
            output = model(input_ids, input_mask)
            
            # Move predictions to CPU and store
            predictions = output.argmax(dim=1).detach().cpu().tolist()
            tweet_ids = tweet_id.tolist()
            
            for id, pred in zip(tweet_ids, predictions):
                y_pred[id] = pred

df1 = pd.DataFrame(y_pred.items())

df1.columns = ['tweet_id', 'Class']

df1['Class'].value_counts()

positive_df = df1[df1['Class'] == 1]

positive_df.sample(5)

df[df['tweet_id'].isin(positive_df['tweet_id'].tolist())]

df1.to_csv('prediction_task1.tsv',sep='\t',index=False)
